Create ArduinoCode
This code reads the current position from the motor encoder (pos), then compares that to a desired position (desiredPos) to generate an error e. The derivative and integral of the error are found and these are fed into a PID function (Kp*e + Kd*de + Ki*inte) which is converted from a voltage value to a PWM value to send to an Arduino pin which drives the motor. The Serial interface is constantly awaiting the input of any character at which point it will step the desiredPosition up by 90 degrees. In future implementation for the MiniProject, the desiredPos variable will need to be interpreted from a signal that is sent over I2C from the Pi to the Arduino.